+++ src/uvmsc/reg/uvm_reg_fifo.cpp
@@ -97,7 +97,7 @@ void uvm_reg_fifo::set_compare( uvm_check_e check )
 
 unsigned int uvm_reg_fifo::size()
 {
-  return fifo.size();
+  return (int)fifo.size();
 }
 
 
@@ -191,7 +191,7 @@ void uvm_reg_fifo::update( uvm_status_e&      status,
 
   m_update_in_progress = true;
 
-  for( int i = fifo.size() - m_set_cnt;
+  for( int i = (int)fifo.size() - m_set_cnt;
        m_set_cnt > 0;
        i++, m_set_cnt--)
   {
