+++ src/uvmsc/reg/uvm_vreg.cpp
@@ -1472,7 +1472,7 @@ void uvm_vreg::add_field( uvm_vreg_field* field )
   if (idx >= m_fields.size())
   {
     m_fields.push_back(field);
-    idx = m_fields.size()-1;
+    idx = (int)m_fields.size()-1;
   }
 
   m_n_used_bits += field->get_n_bits();
